requirements.txt
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
pyperclip>=1.8.2
